// packages/frontend/src/types/schemas/prisma-generated/inputTypeSchemas/BookmarkUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';

export const BookmarkUncheckedCreateInputSchema: z.ZodType<Prisma.BookmarkUncheckedCreateInput> = z.object({
  id: z.string().uuid().optional(),
  userId: z.string(),
  projectId: z.string(),
  createdAt: z.coerce.date().optional()
}).strict();

export default BookmarkUncheckedCreateInputSchema;